Removed CityInput file and combined into CitySearch. It made more sense to have it all in one component as its all doing the same thing. Also added next and search button to keyboard, working properly.
